# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    IQmeetEQ Workshop Companion API.

    Cohort-based workshop app: each workshop runs as its own virtual app
    instance keyed by cohort code. Participants sign in with name + email +
    cohort code; admins sign in with a single admin password.

    All participant endpoints require an authenticated session cookie
    (`iqmeq_session`). Admin endpoints additionally require an admin session.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: participantLogin
      tags: [auth]
      summary: Log in (or sign up) a participant by name + email + cohort code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/check-email:
    post:
      operationId: participantCheckEmail
      tags: [auth]
      summary: Check whether an email is already registered as a participant
      description: Lightweight existence check that does NOT create or modify any session. Used by the login form to switch into the "returning participant" flow.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckEmailRequest" }
      responses:
        "200":
          description: Lookup result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckEmailResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /auth/me:
    get:
      operationId: getCurrentParticipant
      tags: [auth]
      summary: Return the currently authenticated participant + cohort
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /sections:
    get:
      operationId: listSections
      tags: [sections]
      summary: List visible sections for the current participant's cohort
      responses:
        "200":
          description: Sections
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SectionsResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /sections/unlock:
    post:
      operationId: unlockSection
      tags: [sections]
      summary: Submit an unlock code; unlocks all matching sections in the cohort
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UnlockRequest" }
      responses:
        "200":
          description: Unlocked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnlockResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    LoginRequest:
      type: object
      required: [name, email]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        # Optional: when omitted, the server resolves the participant by
        # email alone (their most recent cohort). The login form sends an
        # empty/omitted code in the returning-user flow once the email is
        # recognized via /auth/check-email.
        cohortCode: { type: string }
    CheckEmailRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    CheckEmailResponse:
      type: object
      required: [exists]
      properties:
        exists: { type: boolean }
        name:
          type: string
          nullable: true
    Participant:
      type: object
      required: [id, name, email, cohortId, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        cohortId: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    Cohort:
      type: object
      required: [id, name, facilitatorMessage, tierAccess, audienceType]
      properties:
        id: { type: integer }
        name: { type: string }
        facilitatorMessage: { type: string }
        tierAccess:
          type: object
          additionalProperties: { type: boolean }
          description: 'Map of tier level to default-unlocked boolean, e.g. {"1": true, "2": false}'
        audienceType: { type: string }
    LoginResponse:
      type: object
      required: [participant, isNew]
      properties:
        participant: { $ref: "#/components/schemas/Participant" }
        isNew: { type: boolean }
    MeResponse:
      type: object
      required: [participant, cohort]
      properties:
        participant: { $ref: "#/components/schemas/Participant" }
        cohort: { $ref: "#/components/schemas/Cohort" }
...
    Section:
      type: object
      required: [id, title, description, level, sortOrder, type, unlocked, hasCode, isGeneric]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        level: { type: integer }
        sortOrder: { type: integer }
        type:
          type: string
          enum: [exercise, reference, locked]
        unlocked: { type: boolean }
        hasCode: { type: boolean }
        isGeneric: { type: boolean }
        generic:
          nullable: true
          type: object
          required: [content, promptBlock, goalText]
          properties:
            content: { type: string }
            promptBlock: { type: string, nullable: true }
            goalText: { type: string, nullable: true }
    SectionsResponse:
      type: object
      required: [sections]
      properties:
        sections:
          type: array
          items: { $ref: "#/components/schemas/Section" }
    UnlockRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 1 }
    UnlockedSectionSummary:
      type: object
      required: [id, title, level, type]
      properties:
        id: { type: string }
        title: { type: string }
        level: { type: integer }
        type: { type: string }
    UnlockResponse:
      type: object
      required: [section, unlockedCount]
      properties:
        section: { $ref: "#/components/schemas/UnlockedSectionSummary" }
        unlockedCount: { type: integer }
...