# lib/api-zod/src/generated/types/checkEmailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQmeetEQ Workshop Companion API.

Cohort-based workshop app: each workshop runs as its own virtual app
instance keyed by cohort code. Participants sign in with name + email +
cohort code; admins sign in with a single admin password.

All participant endpoints require an authenticated session cookie
(`iqmeq_session`). Admin endpoints additionally require an admin session.

 * OpenAPI spec version: 0.1.0
 */

export interface CheckEmailResponse {
  exists: boolean;
  name?: string | null;
}